Moved maintaining of sufficient statistics to their own class.
Cleaned up output of printed convergence statistics.

Removed the check for convergence within an epoch -- this
usually just results in doing fewer batches per epoch, it would
be better to look for epochs converging.

Moved small functions to a utilities file.

Changed e-step to save sufficient statistics as an option, this is
needed for when doing fit-predict method where you don't want the
sufficient statistics to be modified because you're evaluating
performance on the test set.

Implemented the option to use smoothed gradients as described in
Mandt and Blei's Smoothed Gradients for Stochastic Variational Inference
paper.